fix: keep Wire package within file-count boundary
