+++ pkg/wire/factory_definition_packaged_installation_test.go
@@ -1,11 +1,13 @@
 package wire
 
 import (
+	"bytes"
 	"io/fs"
 	"path/filepath"
 	"testing"
 	"time"
 
+	packagedfactories "github.com/portpowered/infinite-you/packages/packaged-factories"
 	platformclock "github.com/portpowered/infinite-you/pkg/platform/clock"
 	platformfilesystem "github.com/portpowered/infinite-you/pkg/platform/filesystem"
 	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
@@ -89,5 +91,50 @@ func TestProvideFactoryDefinitionRemainingEffectsSelectOverridesOrPolicyFreeDefa
 	}
 	if got := provideFactoryDefinitionPortableBundledFileInspection(serviceedges.Edges{}); got != (platformfilesystem.Local{}) {
 		t.Fatalf("portable inspection default = %T, want platform filesystem Local", got)
+	}
+}
+
+func TestProvidePackagedFactoryDefinitions_LoadsDetachedGeneratedCatalog(t *testing.T) {
+	definitions, err := providePackagedFactoryDefinitions()
+	if err != nil {
+		t.Fatalf("providePackagedFactoryDefinitions() error = %v", err)
+	}
+
+	wantNames := []string{
+		"@you/deep-research",
+		"@you/fusion",
+		"@you/goal",
+		"@you/quorum",
+		"@you/review",
+		"@you/subagent",
+		"@you/tts",
+	}
+	if len(definitions) != len(wantNames) {
+		t.Fatalf("definition count = %d, want %d", len(definitions), len(wantNames))
+	}
+	for index, wantName := range wantNames {
+		if definitions[index].Name != wantName {
+			t.Fatalf("definitions[%d].Name = %q, want %q", index, definitions[index].Name, wantName)
+		}
+	}
+
+	publishedGoal, err := fs.ReadFile(
+		packagedfactories.Published(),
+		"generated/factories/goal/factory.json",
+	)
+	if err != nil {
+		t.Fatalf("read published Goal definition: %v", err)
+	}
+	if !bytes.Equal(definitions[2].JSON, publishedGoal) {
+		t.Fatal("injected Goal definition differs from the generated publication artifact")
+	}
+
+	definitions[2].JSON[0] ^= 0xff
+	reloaded, err := providePackagedFactoryDefinitions()
+	if err != nil {
+		t.Fatalf("second providePackagedFactoryDefinitions() error = %v", err)
+	}
+	if !bytes.Equal(reloaded[2].JSON, publishedGoal) {
+		t.Fatal("mutating one injected catalog changed a later injected catalog")
 	}
 }
